pyproject.toml
[project]
name = "memenote"
version = "0.1.0"
description = "a Fastapi REST API for memenote"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.14.1",
    "argon2-cffi>=23.1.0",
    "celery>=5.4.0",
    "fastapi[standard]>=0.115.9",
    "passlib>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pyjwt>=2.10.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.38",
    "sse-starlette>=2.2.1",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
